Check for window object when using
Now that the code will be executed in an environment where window might
not exist, we should check to make sure it is available before trying
to use it.

+++ src/entry.js
@@ -16,7 +16,7 @@ const render = () => {
   )
 }
 
-window.localStorage.clear()
+if (window && window.localStorage) window.localStorage.clear()
 
 render()
 store.subscribe(render)
